+++ core/execution/engine_snapshot.go
@@ -15,6 +15,7 @@ package execution
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"code.vegaprotocol.io/vega/core/types"
 	"code.vegaprotocol.io/vega/libs/proto"
@@ -34,6 +35,9 @@ func (e *Engine) marketsStates() ([]*types.ExecMarket, []types.StateProvider) {
 	}
 	e.newGeneratedProviders = make([]types.StateProvider, 0, mkts*4)
 	for _, m := range e.marketsCpy {
+		// ensure the next MTM timestamp is set correctly:
+		am := e.markets[m.mkt.ID]
+		m.nextMTM = am.nextMTM
 		e.log.Debug("serialising market", logging.String("id", m.mkt.ID))
 		mks = append(mks, m.getState())
 
@@ -75,6 +79,7 @@ func (e *Engine) restoreMarket(ctx context.Context, em *types.ExecMarket) (*Mark
 		return nil, err
 	}
 
+	nextMTM := time.Unix(0, em.NextMTM)
 	// create market auction state
 	e.log.Info("restoring market", logging.String("id", em.Market.ID))
 	mkt, err := NewMarketFromSnapshot(
@@ -110,6 +115,8 @@ func (e *Engine) restoreMarket(ctx context.Context, em *types.ExecMarket) (*Mark
 	if err := e.propagateInitialNetParams(ctx, mkt); err != nil {
 		return nil, err
 	}
+	// ensure this is set correctly
+	mkt.nextMTM = nextMTM
 
 	e.publishNewMarketInfos(ctx, mkt)
 	return mkt, nil