+++ core/execution/market_test.go
@@ -411,6 +411,7 @@ func getTestMarket2WithDP(
 	}
 	require.NoError(t, err)
 	mkt := getMarketWithDP(pMonitorSettings, openingAuctionDuration, decimalPlaces)
+	// ensure nextMTM is happening every block
 	mktCfg := &mkt
 	mktCfg.DecimalPlaces = cfgAsset.DecimalPlaces()
 
@@ -428,6 +429,8 @@ func getTestMarket2WithDP(
 	if err != nil {
 		t.Fatalf("couldn't create a market: %v", err)
 	}
+	// ensure MTM settlements happen every block
+	mktEngine.OnMarkPriceUpdateMaximumFrequency(context.Background(), time.Duration(0))
 	mktEngine.UpdateRiskFactorsForTest()
 
 	if startOpeningAuction {
@@ -1837,6 +1840,7 @@ func TestTriggerByPriceNoTradesInAuction(t *testing.T) {
 		Reference:   "party2-sell-order-2",
 	}
 	confirmationSell, err = tm.market.SubmitOrder(context.Background(), orderSell2)
+	tm.market.OnTick(vegacontext.WithTraceID(context.Background(), vgcrypto.RandomHash()), now)
 	require.NotNil(t, confirmationSell)
 	require.NoError(t, err)
 	require.Equal(t, types.MarketStateSuspended, tm.market.State()) // enter auction
@@ -6390,6 +6394,8 @@ func Test3008And3007CancelLiquidityProvision(t *testing.T) {
 	cnf, err := tm.market.SubmitOrder(ctx, newOrder)
 	assert.NoError(t, err)
 	assert.True(t, len(cnf.Trades) > 0)
+	// force MTM transfers here, we reset the events after
+	tm.market.OnTick(ctx, tm.now)
 
 	// clean the events
 	// then check for transfer of liquidity fees
@@ -6814,10 +6820,11 @@ func Test3045DistributeFeesToManyProviders(t *testing.T) {
 		TimeInForce: types.OrderTimeInForceGTC,
 	})
 
-	tm.events = nil
 	cnf, err := tm.market.SubmitOrder(ctx, newOrder)
 	assert.NoError(t, err)
 	assert.True(t, len(cnf.Trades) > 0)
+	// force MTM
+	tm.market.OnTick(ctx, tm.now)
 
 	// clean the events
 	// then check for transfer of liquidity fees