+++ datanode/gateway/graphql/generated.go
@@ -756,6 +756,7 @@ type ComplexityRoot struct {
 		MarketTradingMode         func(childComplexity int) int
 		MarketValueProxy          func(childComplexity int) int
 		MidPrice                  func(childComplexity int) int
+		NextMarkToMarket          func(childComplexity int) int
 		OpenInterest              func(childComplexity int) int
 		PriceMonitoringBounds     func(childComplexity int) int
 		StaticMidPrice            func(childComplexity int) int
@@ -961,6 +962,7 @@ type ComplexityRoot struct {
 		MarketTradingMode         func(childComplexity int) int
 		MarketValueProxy          func(childComplexity int) int
 		MidPrice                  func(childComplexity int) int
+		NextMarkToMarket          func(childComplexity int) int
 		OpenInterest              func(childComplexity int) int
 		PriceMonitoringBounds     func(childComplexity int) int
 		StaticMidPrice            func(childComplexity int) int
@@ -1887,6 +1889,7 @@ type MarketDataResolver interface {
 	PriceMonitoringBounds(ctx context.Context, obj *vega.MarketData) ([]*PriceMonitoringBounds, error)
 
 	LiquidityProviderFeeShare(ctx context.Context, obj *vega.MarketData) ([]*LiquidityProviderFeeShare, error)
+	NextMarkToMarket(ctx context.Context, obj *vega.MarketData) (string, error)
 }
 type MarketDepthResolver interface {
 	Market(ctx context.Context, obj *vega.MarketDepth) (*vega.Market, error)
@@ -1948,6 +1951,7 @@ type ObservableMarketDataResolver interface {
 	PriceMonitoringBounds(ctx context.Context, obj *vega.MarketData) ([]*PriceMonitoringBounds, error)
 
 	LiquidityProviderFeeShare(ctx context.Context, obj *vega.MarketData) ([]*ObservableLiquidityProviderFeeShare, error)
+	NextMarkToMarket(ctx context.Context, obj *vega.MarketData) (string, error)
 }
 type ObservableMarketDepthResolver interface {
 	LastTrade(ctx context.Context, obj *vega.MarketDepth) (*MarketDepthTrade, error)
@@ -4783,6 +4787,13 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 
 		return e.complexity.MarketData.MidPrice(childComplexity), true
 
+	case "MarketData.nextMarkToMarket":
+		if e.complexity.MarketData.NextMarkToMarket == nil {
+			break
+		}
+
+		return e.complexity.MarketData.NextMarkToMarket(childComplexity), true
+
 	case "MarketData.openInterest":
 		if e.complexity.MarketData.OpenInterest == nil {
 			break
@@ -5655,6 +5666,13 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 		}
 
 		return e.complexity.ObservableMarketData.MidPrice(childComplexity), true
+
+	case "ObservableMarketData.nextMarkToMarket":
+		if e.complexity.ObservableMarketData.NextMarkToMarket == nil {
+			break
+		}
+
+		return e.complexity.ObservableMarketData.NextMarkToMarket(childComplexity), true
 
 	case "ObservableMarketData.openInterest":
 		if e.complexity.ObservableMarketData.OpenInterest == nil {
@@ -27270,6 +27288,8 @@ func (ec *executionContext) fieldContext_Market_data(ctx context.Context, field 
 				return ec.fieldContext_MarketData_marketValueProxy(ctx, field)
 			case "liquidityProviderFeeShare":
 				return ec.fieldContext_MarketData_liquidityProviderFeeShare(ctx, field)
+			case "nextMarkToMarket":
+				return ec.fieldContext_MarketData_nextMarkToMarket(ctx, field)
 			}
 			return nil, fmt.Errorf("no field named %q was found under type MarketData", field.Name)
 		},
@@ -28822,6 +28842,50 @@ func (ec *executionContext) fieldContext_MarketData_liquidityProviderFeeShare(ct
 	return fc, nil
 }
 
+func (ec *executionContext) _MarketData_nextMarkToMarket(ctx context.Context, field graphql.CollectedField, obj *vega.MarketData) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_MarketData_nextMarkToMarket(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return ec.resolvers.MarketData().NextMarkToMarket(rctx, obj)
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_MarketData_nextMarkToMarket(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "MarketData",
+		Field:      field,
+		IsMethod:   true,
+		IsResolver: true,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			return nil, errors.New("field of type String does not have child fields")
+		},
+	}
+	return fc, nil
+}
+
 func (ec *executionContext) _MarketDataCommitments_sells(ctx context.Context, field graphql.CollectedField, obj *MarketDataCommitments) (ret graphql.Marshaler) {
 	fc, err := ec.fieldContext_MarketDataCommitments_sells(ctx, field)
 	if err != nil {
@@ -29109,6 +29173,8 @@ func (ec *executionContext) fieldContext_MarketDataEdge_node(ctx context.Context
 				return ec.fieldContext_MarketData_marketValueProxy(ctx, field)
 			case "liquidityProviderFeeShare":
 				return ec.fieldContext_MarketData_liquidityProviderFeeShare(ctx, field)
+			case "nextMarkToMarket":
+				return ec.fieldContext_MarketData_nextMarkToMarket(ctx, field)
 			}
 			return nil, fmt.Errorf("no field named %q was found under type MarketData", field.Name)
 		},
@@ -34657,6 +34723,50 @@ func (ec *executionContext) fieldContext_ObservableMarketData_liquidityProviderF
 				return ec.fieldContext_ObservableLiquidityProviderFeeShare_averageEntryValuation(ctx, field)
 			}
 			return nil, fmt.Errorf("no field named %q was found under type ObservableLiquidityProviderFeeShare", field.Name)
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _ObservableMarketData_nextMarkToMarket(ctx context.Context, field graphql.CollectedField, obj *vega.MarketData) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_ObservableMarketData_nextMarkToMarket(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return ec.resolvers.ObservableMarketData().NextMarkToMarket(rctx, obj)
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_ObservableMarketData_nextMarkToMarket(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "ObservableMarketData",
+		Field:      field,
+		IsMethod:   true,
+		IsResolver: true,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			return nil, errors.New("field of type String does not have child fields")
 		},
 	}
 	return fc, nil
@@ -44262,6 +44372,8 @@ func (ec *executionContext) fieldContext_Query_getMarketDataHistoryByID(ctx cont
 				return ec.fieldContext_MarketData_marketValueProxy(ctx, field)
 			case "liquidityProviderFeeShare":
 				return ec.fieldContext_MarketData_liquidityProviderFeeShare(ctx, field)
+			case "nextMarkToMarket":
+				return ec.fieldContext_MarketData_nextMarkToMarket(ctx, field)
 			}
 			return nil, fmt.Errorf("no field named %q was found under type MarketData", field.Name)
 		},
@@ -51405,6 +51517,8 @@ func (ec *executionContext) fieldContext_Subscription_marketsData(ctx context.Co
 				return ec.fieldContext_ObservableMarketData_marketValueProxy(ctx, field)
 			case "liquidityProviderFeeShare":
 				return ec.fieldContext_ObservableMarketData_liquidityProviderFeeShare(ctx, field)
+			case "nextMarkToMarket":
+				return ec.fieldContext_ObservableMarketData_nextMarkToMarket(ctx, field)
 			}
 			return nil, fmt.Errorf("no field named %q was found under type ObservableMarketData", field.Name)
 		},
@@ -65642,6 +65756,26 @@ func (ec *executionContext) _MarketData(ctx context.Context, sel ast.SelectionSe
 				return innerFunc(ctx)
 
 			})
+		case "nextMarkToMarket":
+			field := field
+
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				defer func() {
+					if r := recover(); r != nil {
+						ec.Error(ctx, ec.Recover(ctx, r))
+					}
+				}()
+				res = ec._MarketData_nextMarkToMarket(ctx, field, obj)
+				if res == graphql.Null {
+					atomic.AddUint32(&invalids, 1)
+				}
+				return res
+			}
+
+			out.Concurrently(i, func() graphql.Marshaler {
+				return innerFunc(ctx)
+
+			})
 		default:
 			panic("unknown field " + strconv.Quote(field.Name))
 		}
@@ -67394,6 +67528,26 @@ func (ec *executionContext) _ObservableMarketData(ctx context.Context, sel ast.S
 					}
 				}()
 				res = ec._ObservableMarketData_liquidityProviderFeeShare(ctx, field, obj)
+				return res
+			}
+
+			out.Concurrently(i, func() graphql.Marshaler {
+				return innerFunc(ctx)
+
+			})
+		case "nextMarkToMarket":
+			field := field
+
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				defer func() {
+					if r := recover(); r != nil {
+						ec.Error(ctx, ec.Recover(ctx, r))
+					}
+				}()
+				res = ec._ObservableMarketData_nextMarkToMarket(ctx, field, obj)
+				if res == graphql.Null {
+					atomic.AddUint32(&invalids, 1)
+				}
 				return res
 			}